fix currency dropdown not showing, rename DropdownCurrency component to CurrencyDropdown
